Initial commit. Bare minimum changes to support SO2 sensor.
1c721dd0bb5db563ec03240dd9ff0bece87f1f4f started the process, but did not follow through. `analysis_all_data()` got entries for the SO2 and HCL sensors, but `read_gas_concentration()` did not (the code is effectively replicated in both places). Added this with the minimum amount of changes to get it working for now; will clean up later.

Added `tempSwitch` variable to the `DFRobot_MultiGasSensor` class and removed the weird module-global variable. The `set_temp_compensation()` method now sets the flag specific to that instance, which should keep things more consistent.

Removed `temp` variable from module. The `DFRobot_MultiGasSensor` class already has a `temp` variable, so using that where possible.

If `tempSwitch` is set on an instance, then `read_gas_concentration()` will update the current temperature for the instance before generating the sensor measurement. This would seem to make far more sense than using the temperature that was read once when it was turned on with `set_temp_compensation()`.
￼
Note: there seem to be a host of errors associated in the handling of the thresholds for temperature compensation. I'm not fixing those just yet.