2e625474c39c923b4cee9cf9075cb9021c0f18a3 incorrectly stated that `read_gas_concentration()` updated the temperature prior to doing temperature compenstaion if set. This change does it for realsies.
Also adds one more tweak to deal with the removal of the `temp` variable at the module level. `analysis_all_data()` wants a temperature, so it should use the last temperature taken for the instance. Not entirely sure what this function is for at the moment, but there you go.