Added units to `DFRobot_MultiGasSensor` class. Reading the sensor will result in the units being set appropriately for the sensor type, resulting in application code not needing to keep track of which type of sensor reports what. Granted, this means that pretty much all of the sensors will report "ppm" as only the O2 sensor reports differently (%), but it's the thought that counts.
Trying to consistently use the `gasconcentration` class variable as the raw sensor reading so that we shouldn't be able to get to the point where we try to temperature-correct an already corrected measurement.

Some minor formatting tweaks to the output from the example `read_gas_concentration.py` code. Now displays temperature alongside the gas measurement, and restricts the number of decimal places shown to something saner. Tried to keep things good for python 2, so no f-strings.